Tabs Refactor (#491)
* tabs refactor
* added docstrings and removed caps from import labels
